LayoutTests/editing/smart-lists/ordered-markers-with-different-delimiters-generate-list.html

<!DOCTYPE html><!-- webkit-test-runner [ SmartListsAvailable=true ] -->
<html>
<head>
<script src="../editing.js"></script>
<script src="../../resources/dump-as-markup.js"></script>
</head>
<body>
<div id="test" contenteditable></div>
<script>
Markup.description('Ordered-list markers on consecutive lines should be treated as compatible even when they use different delimiters (period and parenthesis), producing a single ordered list.');

var test = document.getElementById('test');
test.focus();
[...'1. A\n2) B'].forEach(typeCharacterCommand);
Markup.dump('test');
</script>
</body>
</html>
